requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.1.2
Flask==1.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pycparser==2.21
SQLAlchemy==1.4.37
typing-extensions==4.1.1
Werkzeug==0.15.6
WTForms==3.0.1
zipp==3.8.0
